LRCI-6840 Move JobHistory into its own class heirarchy & package
