+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BatchHistory.java
@@ -5,6 +5,8 @@
 
 package com.liferay.jenkins.results.parser;
 
+import com.liferay.jenkins.results.parser.history.JobHistory;
+
 import java.util.HashMap;
 import java.util.Map;
 
@@ -16,27 +18,7 @@ import org.json.JSONObject;
  */
 public class BatchHistory {
 
-	public long getAverageDuration() {
-		return _averageDuration;
-	}
-
-	public String getBatchName() {
-		return _batchName;
-	}
-
-	public JobHistory getJobHistory() {
-		return _jobHistory;
-	}
-
-	public TestHistory getTestHistory(String key) {
-		return _testHistories.get(key);
-	}
-
-	public TestTaskHistory getTestTaskHistory(String key) {
-		return _testTaskHistories.get(key);
-	}
-
-	protected BatchHistory(JobHistory jobHistory, JSONObject jsonObject) {
+	public BatchHistory(JobHistory jobHistory, JSONObject jsonObject) {
 		_jobHistory = jobHistory;
 
 		_averageDuration = jsonObject.optLong("averageDuration");
@@ -64,6 +46,26 @@ public class BatchHistory {
 					testTaskHistory.getTestTaskName(), testTaskHistory);
 			}
 		}
+	}
+
+	public long getAverageDuration() {
+		return _averageDuration;
+	}
+
+	public String getBatchName() {
+		return _batchName;
+	}
+
+	public JobHistory getJobHistory() {
+		return _jobHistory;
+	}
+
+	public TestHistory getTestHistory(String key) {
+		return _testHistories.get(key);
+	}
+
+	public TestTaskHistory getTestTaskHistory(String key) {
+		return _testTaskHistories.get(key);
 	}
 
 	private final long _averageDuration;
