+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TestHistory.java
@@ -5,6 +5,8 @@
 
 package com.liferay.jenkins.results.parser;
 
+import com.liferay.jenkins.results.parser.history.JobHistory;
+
 import org.json.JSONObject;
 
 /**
